requirements.txt
discord
google-cloud-translate
python-dotenv
